+++ packages/interactive_media_ads/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.2+13
+
+* Adds internal wrapper for Android native `Ad`.
+
+## 0.2.2+12
+
+* Adds internal wrapper for iOS native `IMACompanionAd`.
+
 ## 0.2.2+11
 
 * Adds internal wrapper for Android native `UniversalAdId`.